Squashed commit of the following:
commit 78793b20816ea43fcb4a4726f60e64c3ccb9de95
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Sat Jun 24 04:45:53 2023 +0800

    Minor follow-up fixes.

commit db563a7b37a7eedca4c60fb540ee34540967990e
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Sat Jun 24 04:19:17 2023 +0800

    Revert back documents setup.

commit 5f8af2dbe8c92ce7bda307074518d94e4386b0c8
Merge: 22ae572 2bdd951
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Sat Jun 24 04:14:05 2023 +0800

    Merge branch 'development' of https://github.com/ddreilll/pupqc-srims into development

commit 22ae572847ec9e87709a5569b44d0bc939a26f08
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Sat Jun 24 04:12:53 2023 +0800

    Major Updates

commit 2bdd95135df664a9923303a07ed2686cd4265755
Author: John Chris Cayetano <cayetanojohnchris@gmail.com>
Date:   Sat Jun 24 00:17:13 2023 +0800

    Update

commit 55435c7690e3e58f88b8ab74ed7fc35bdb626118
Author: John Chris Cayetano <cayetanojohnchris@gmail.com>
Date:   Fri Jun 23 22:50:10 2023 +0800

    Lang implementation at Menu and Settings Index

commit 76749d7b42962ccd62fd9cb28aae62ffd22b1d61
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Fri Jun 23 21:24:27 2023 +0800

    System Setup UI Enhancements

commit 9d06114861b59f9cbd1db8eaa7d464a59634e3d5
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Fri Jun 23 15:44:47 2023 +0800

    Permission fixed.

commit c55f54362a9e738b87852d89d22f71bdde2a8a1d
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Fri Jun 23 00:13:42 2023 +0800

    Delete SubjectController.temp

commit c329e1616272c6d348f63963143a4eb1882f19d4
Merge: 5a51d38 6e2497f
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Fri Jun 23 00:13:11 2023 +0800

    Merge branch 'development' of https://github.com/ddreilll/pupqc-srims into development

commit 5a51d3882626082bb1a6665cf37377957ae874e4
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Fri Jun 23 00:12:26 2023 +0800

    Enhanced instructor setup

commit 6e2497f8149d4456e9a04c42e4d68ff085c2037f
Author: John Chris Cayetano <cayetanojohnchris@gmail.com>
Date:   Fri Jun 23 00:11:39 2023 +0800

    Subject CRUD

    Views

commit 3cfcef6046baccfcca79b32d75f2ff491b19eb36
Author: John Chris Cayetano <cayetanojohnchris@gmail.com>
Date:   Thu Jun 22 23:48:23 2023 +0800

    Subject CRUD

    Subject's Controller. Model. Request, and Route.

commit 895de33df309ff1beb7f32f4e53cc02bb4752f1e
Merge: 498a228 b17033a
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Thu Jun 22 23:08:35 2023 +0800

    Merge branch 'development' of https://github.com/ddreilll/pupqc-srims into development

commit 498a228890abed46dda3c1406cce3bf27cb992d8
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Thu Jun 22 23:08:31 2023 +0800

    System Setup updates

commit b17033a41eb33d1b8050625eeb1b2a769a8490dc
Author: John Chris Cayetano <cayetanojohnchris@gmail.com>
Date:   Thu Jun 22 23:07:45 2023 +0800

    CRUDS Lang

commit 998e35e748ebd4f8506686327dddca60020da29c
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Thu Jun 22 21:25:06 2023 +0800

    Enhanced document setup

commit 30a03e47fbc54115c2c14c0801a640fccba61efa
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Thu Jun 22 17:40:19 2023 +0800

    Urgent fix on web.php

commit a63d2b989ea4477bbdc8e831f71a686a89058ef5
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Thu Jun 22 13:52:59 2023 +0800

    Fixed Profile edit

commit 4fcb194f7b6e44489b2b63dfeb8069754b0fed23
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Thu Jun 22 13:45:04 2023 +0800

    Bug Fixes

commit d22237ccf65b710feb5008da6d5b33b8fe2450ac
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Thu Jun 22 13:20:48 2023 +0800

    Minor bug fixes

commit c3586b82a8beba8050b7c8f7cb959632ccddb1a0
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Thu Jun 22 11:07:18 2023 +0800

    Added DB backup file

commit b50ee34498dd3c71fc685185d117615fa6da3f5d
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Thu Jun 22 11:01:11 2023 +0800

    Added README

commit 0eed67b2efa9183c70e1383a2f7c0b57b36d56f2
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Thu Jun 22 10:53:47 2023 +0800

    Minor changes

    Initialized UI re-arrangement on System Settings

commit 4af65633b68ed2dd98850416fbc8afa6ed3f9819
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Thu Jun 22 10:32:51 2023 +0800

    Enhanced System Settings | Code cleanup | Bug Fixes

commit 4fdb62ed1ba35c6875f2a3a8a2adb25993a330db
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Wed Jun 21 02:06:42 2023 +0800

    Update User Management actions button

commit 2c36f49619dda06d55209293ee88e6f72887428e
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Wed Jun 21 01:56:59 2023 +0800

    Added User Account Management | UI Enhancements

commit 6b178967a2295e938b07d1ad5283232151fd4dc6
Merge: bc931bf c4530dc
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Tue Jun 20 18:48:28 2023 +0800

    Merge branch 'stable' into development

commit bc931bf9f5c038e26c14c1308d99ddfc4ec8c7f9
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Tue Jun 20 16:51:50 2023 +0800

    Fixed Document Evaluation Report

commit e03885ce03b42b493b03064f492301475a337555
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Tue Jun 20 16:24:33 2023 +0800

    Added Gradesheet generator | Bug Fixes | Code cleanup

    - Added Gradesheet PDF Generator
    - Fixed Student Scholastic Report Generator
    - Code cleanup & Re-organization

commit 51363acb917e011440b8fc7d4c3168c66fcb9a92
Author: J-Cayetano <cayetanojohnchris@gmail.com>
Date:   Tue Jun 20 14:18:51 2023 +0800

    Student Scholastic Record (PDF)

    Generation of Scholastic Record PDF per Students

+++ app/Http/Controllers/Admin/HonorsController.php
@@ -4,15 +4,14 @@ namespace App\Http\Controllers\Admin;
 
 use App\Models\Honor;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\StoreHonorRequest;
 use App\Http\Requests\UpdateHonorRequest;
 
 class HonorsController extends Controller
 {
-
-    protected $menuHeader = "System Settings";
-    protected $menu = "system-settings";
 
     /**
      * Display a listing of the resource.
@@ -21,22 +20,11 @@ class HonorsController extends Controller
      */
     public function index()
     {
-        // abort_if(Gate::denies('honor_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
-
-        $menuHeader = $this->menuHeader;
-        $menu = $this->menu;
-
-
-        $title = "Honors";
-        $breadcrumb = [
-            [
-                "name" => "Honors"
-            ]
-        ];
+        abort_if(Gate::denies('honor_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $honors = Honor::all();
 
-        return view('admin.honors.index', compact('menuHeader', 'title', 'menu', 'breadcrumb', 'honors'));
+        return view('admin.honors.index', compact('honors'));
     }
 
     /**
@@ -46,24 +34,9 @@ class HonorsController extends Controller
      */
     public function create()
     {
-        // abort_if(Gate::denies('honor_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('honor_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
-        $menuHeader = $this->menuHeader;
-        $menu = $this->menu;
-
-
-        $title = "Add Honor";
-        $breadcrumb = [
-            [
-                "name" => "Honors",
-                "url" => route('settings.honors.index')
-            ],
-            [
-                "name" => "Add Honor"
-            ]
-        ];
-
-        return view('admin.honors.create', compact('menuHeader', 'title', 'menu', 'breadcrumb'));
+        return view('admin.honors.create');
     }
 
     /**
@@ -88,24 +61,9 @@ class HonorsController extends Controller
      */
     public function edit(Honor $honor)
     {
-        // abort_if(Gate::denies('honor_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('honor_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
-        $menuHeader = $this->menuHeader;
-        $menu = $this->menu;
-
-
-        $title = "Edit Honor";
-        $breadcrumb = [
-            [
-                "name" => "Honors",
-                "url" => route('settings.honors.index')
-            ],
-            [
-                "name" => "Edit Honor"
-            ]
-        ];
-
-        return view('admin.honors.edit', compact('menuHeader', 'title', 'menu', 'breadcrumb', 'honor'));
+        return view('admin.honors.edit', compact('honor'));
     }
 
     /**
@@ -132,7 +90,7 @@ class HonorsController extends Controller
      */
     public function destroy(Honor $honor)
     {
-        // abort_if(Gate::denies('honor_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('honor_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $honor->delete();
